mv context public APIs to separate file
