+++ core/frontend/src/common/render/GraphicDescriptionBuilder.ts
@@ -7,13 +7,12 @@
  */
 
 import { GraphicAssembler } from "./GraphicAssembler";
-import { WorkerGraphicDescriptionContextImpl } from "../internal/render/GraphicDescriptionContextImpl";
 import { collectGraphicDescriptionTransferables, GraphicDescriptionBuilderImpl } from "../internal/render/GraphicDescriptionBuilderImpl";
 import { Point3d, Range3d, Transform } from "@itwin/core-geometry";
 import { GraphicType } from "./GraphicType";
 import { PickableGraphicOptions } from "./BatchOptions";
 import { _implementationProhibited } from "../internal/Symbols";
-import { TransientIdSequence } from "@itwin/core-bentley";
+import { WorkerGraphicDescriptionContext } from "./GraphicDescriptionContext";
 
 /** An opaque representation of a [[RenderGraphic]] created by a [[GraphicDescriptionBuilder]].
  * Unlike `RenderGraphic`, a `GraphicDescription` does not allocate any WebGL resources like textures, vertex buffers, etc, so
@@ -35,66 +34,6 @@ export namespace GraphicDescription {
   export function collectTransferables(transferables: Set<Transferable>, description: GraphicDescription): void {
     return collectGraphicDescriptionTransferables(transferables, description);
   }
-}
-
-/** An opaque representation of a [[WorkerGraphicDescriptionContext]] that can be passed from the main thread to a [Worker](https://developer.mozilla.org/en-US/docs/Web/API/Worker).
- * @see [[RenderSystem.createWorkerGraphicDescriptionContextProps]] to obtain an implementation of this type.
- * @see [[WorkerGraphicDescriptionContext.fromProps]] to instantiate the context on a Worker from this representation.
- * @beta
- */
-export interface WorkerGraphicDescriptionContextProps {
-  /** @internal */
-  readonly [_implementationProhibited]: unknown;
-}
-
-/** Context allocated on a [Worker](https://developer.mozilla.org/en-US/docs/Web/API/Worker) to enable it to create [[GraphicDescription]]s.
- * When the Worker returns one or more GraphicDescriptions to the main thread, it should also return this context as a [[GraphicDescriptionContextProps]].
- * @see [[WorkerGraphicDescriptionContext.fromProps]] to instantiate this type.
- * @beta
- */
-export interface WorkerGraphicDescriptionContext {
-  /** @internal */
-  readonly [_implementationProhibited]: unknown;
-  /** A "fork" of the iModel's transient Id sequence obtained from the main thread. New Ids generated by this sequence on the Worker will be
-   * merged into the original sequence by [[RenderSystem.resolveGraphicDescriptionContext]].
-   */
-  readonly transientIds: TransientIdSequence;
-  /** Converts this context to a representation that can be passed back to the main thread.
-   * @param transferables A set of [transferable objects](https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Transferable_objects) that
-   * can be efficiently copied to the main thread.
-   */
-  toProps(transferables: Set<Transferable>): GraphicDescriptionContextProps;
-}
-
-/** @beta */
-export namespace WorkerGraphicDescriptionContext {
-  /** Instantiate a context from its opaque representation. */
-  export function fromProps(props: WorkerGraphicDescriptionContextProps): WorkerGraphicDescriptionContext {
-    return new WorkerGraphicDescriptionContextImpl(props);
-  }
-}
-
-/** Describes a [[GraphicDescriptionContext]] returned from a [Worker](https://developer.mozilla.org/en-US/docs/Web/API/Worker) to the main thread, holding resources like
- * transient Ids, textures, and materials that were allocated on the Worker for use by [[GraphicDescription]]s.
- * @see [[WorkerGraphicDescriptionContext.toProps]] to obtain an implementation of this type.
- * @see [[RenderSystem.resolveGraphicDescriptionContext]] to instantiate the context from this representation.
- * @beta
- */
-export interface GraphicDescriptionContextProps {
-  /** @internal */
-  readonly [_implementationProhibited]: unknown;
-}
-
-/** Context holding resources like transient Ids, textures, and materials that were allocated on a [Worker](https://developer.mozilla.org/en-US/docs/Web/API/Worker) by a
- * [[WorkerGraphicDescriptionContext]] for use in [[GraphicDescription]]s. This context must be supplied to [[RenderSystem.createGraphicFromDescription]] when converting a
- * GraphicDescription to a [[RenderGraphic]].
- * @see [[RenderSystem.resolveGraphicDescriptionContext]] to obtain an implementation of this type.
- * @beta
- */
-export interface GraphicDescriptionContext {
-  /** @internal */
-  readonly [_implementationProhibited]: unknown;
-  remapTransientLocalId(sourceLocalId: number): number;
 }
 
 /** Arguments supplied to [[GraphicDescriptionBuilderOptions.computeChordTolerance]] to help compute an appropriate level of detail for the [[GraphicDescription]].
